+++ src/constants.ts
@@ -7,6 +7,7 @@ export class Constants {
     public static readonly CONFIGURATION_CONTRACT_SORT_ORDER = "cosmywasmy.contractSortOrder";
     public static readonly CONFIGURATION_COSMWASM_RESPONSE_VIEW = "cosmywasmy.cosmwasmResponseView";
     public static readonly CONFIGURATION_HISTORY_STORED = "cosmywasmy.maxHistoryStored";
+    public static readonly CONFIGURATION_OPEN_TX_IN_VSCODE = "cosmywasmy.openTxExplorerInVscode";
     
     public static readonly VIEWS_ACCOUNT = "account";
     public static readonly VIEWS_CONTRACT = "contract";