+++ src/helpers/Cosmwasm/Executer.ts
@@ -70,7 +70,12 @@ export class Executer {
             const url = global.workspaceChain.txExplorerLink;
             if (tx && url) {
                 const explorerUrl = url.replace("${txHash}", tx);
-                vscode.window.showInformationMessage(new vscode.MarkdownString("View transaction in explorer - [" + tx + "](" + explorerUrl + ")", true).value);
+                if (Workspace.GetOpenTxInSimpleBrowser()) {
+                    vscode.commands.executeCommand("simpleBrowser.api.open", vscode.Uri.parse(explorerUrl));
+                }
+                else {
+                    vscode.window.showInformationMessage(new vscode.MarkdownString("View transaction in explorer - [" + tx + "](" + explorerUrl + ")", true).value);
+                }
                 return Promise.resolve();
             }
             else {